ML/server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
tensorflow==2.15.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
matplotlib==3.8.2
folium==0.15.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0
pytest==7.4.3
hypothesis==6.92.1
ortools==9.8.3296
googlemaps==4.10.0
azure-maps-route==1.0.0b2
gunicorn==21.2.0